requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
google-search==1.1.1
google-search-results==2.4.2
idna==3.4
lxml==4.9.2
python-dotenv==1.0.0
pytube==15.0.0
requests==2.30.0
soupsieve==2.4.1
tqdm==4.65.0
urllib3==2.0.2
